refactor:  replace 'TextSize' with 'ITextSize' in various components and types for consistency
