+++ src/pages/register/index.tsx
@@ -2,7 +2,7 @@ import { useMediaQuery } from "@inubekit/inubekit";
 import { useLocation, useNavigate } from "react-router-dom";
 import { useEffect } from "react";
 import { RegisterUI } from "./interface";
-import { TextSize } from "@ptypes/components/TextSize";
+import { ITextSize } from "@src/types/components/ITextSize";
 import { registerLabels } from "@config/register/labels/registerLabels";
 import type { ILocationState } from "@ptypes/pages/register/ILocationState";
 import type { UserData } from "@ptypes/hooks/useValidationToken/IUserData";
@@ -11,7 +11,7 @@ const Register = () => {
   const location = useLocation() as { state: ILocationState };
   const navigate = useNavigate();
   const screenMobile = useMediaQuery("(max-width: 768px)");
-  const labelsSizeDifferent: TextSize = screenMobile ? "medium" : "large";
+  const labelsSizeDifferent: ITextSize = screenMobile ? "medium" : "large";
 
   const userData: UserData | undefined = location.state?.userData;
   console.log("Register page userData:", userData);
