+++ src/types/components/ITextSize/index.ts
