+++ src/types/components/login/IDynamicInputs/index.ts
@@ -1,7 +1,7 @@
 import { IFormStepLabels } from '@ptypes/hooks/useTwoStepLoginForm/IFormStepLabels';
 import type { ReactNode } from 'react';
 import { InputStatus } from '@ptypes/components/InputStatus';
-import { ITextSize } from '@ptypes/components/TextSize';
+import { ITextSize } from '@src/types/components/ITextSize';
 
 interface IDynamicInputs {
     labels: IFormStepLabels,
