+++ src/types/components/register/IUserWelcome/index.ts
@@ -1,0 +1,13 @@
+export interface IUserWelcome {
+  userData: {
+    firstName: string;
+    lastName: string;
+    documentType: string;
+    documentNumber: string;
+  };
+  labels: {
+    welcomeMessage: {
+      greeting: string;
+    };
+  };
+}
